Déplace le pré-diagnostic des pages vers le domaine PDF
